remote-repo-server/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
pydantic
PyJWT
python-multipart
cryptography
confluent_kafka